Add pagamentos

+++ src/main/java/saas/hotel/istoepousada/repository/HistoricoRecebidosFuncionarioRepository.java
@@ -53,7 +53,7 @@ public class HistoricoRecebidosFuncionarioRepository {
                       new Funcionario.Nome(
                           rs.getLong("pagamento_desconto_funcionario_id"),
                           rs.getString("pagamento_desconto_funcionario_nome")),
-                      rs.getInt("pagamento_desconto_porcentagem"),
+                      rs.getFloat("pagamento_desconto_porcentagem"),
                       rs.getFloat("pagamento_desconto_valor"),
                       rs.getTimestamp("pagamento_desconto_data_hora_registro").toLocalDateTime()),
                   rs.getString("")),
